// src/dto/mai.dto.ts
import z from "zod";

export const UpdateMaiUserOptionsDto = z.object({
	optionKind: z.number().min(0).max(3).optional(),
	noteSpeed: z.number().min(0).max(37).optional(),
	slideSpeed: z.number().min(0).max(37).optional(),
	touchSpeed: z.number().min(0).max(20).optional(),
	tapDesign: z.number().min(0).max(4).optional(),
	holdDesign: z.number().min(0).max(1).optional(),
	slideDesign: z.number().min(0).max(1).optional(),
	starType: z.number().min(0).max(1).optional(),
	outlineDesign: z.number().min(0).max(16).optional(),
	noteSize: z.number().min(0).max(5).optional(),
	slideSize: z.number().min(0).max(5).optional(),
	touchSize: z.number().min(0).max(5).optional(),
	starRotate: z.number().min(0).max(1).optional(),
	dispCenter: z.number().min(0).max(11).optional(),
	dispChain: z.number().min(0).max(2).optional(),
	dispRate: z.number().min(0).max(7).optional(),
	dispBar: z.number().min(0).max(1).optional(),
	touchEffect: z.number().min(0).max(2).optional(),
	submonitorAchieve: z.number().min(0).max(1).optional(),
	submonitorAppeal: z.number().min(0).max(7).optional(),
	matching: z.number().min(0).max(1).optional(),
	trackSkip: z.number().min(0).max(11).optional(),
	brightness: z.number().min(0).max(4).optional(),
	mirrorMode: z.number().min(0).max(3).optional(),
	dispJudge: z.number().min(0).max(14).optional(),
	dispJudgePos: z.number().min(0).max(5).optional(),
	dispJudgeTouchPos: z.number().min(0).max(2).optional(),
	adjustTiming: z.number().min(0).max(40).optional(),
	judgeTiming: z.number().min(0).max(40).optional(),
	ansVolume: z.number().min(0).max(10).optional(),
	tapHoldVolume: z.number().min(0).max(5).optional(),
	criticalSe: z.number().min(0).max(3).optional(),
	tapSe: z.number().min(0).max(7).optional(),
	breakSe: z.number().min(0).max(49).optional(),
	breakVolume: z.number().min(0).max(5).optional(),
	exSe: z.number().min(0).max(6).optional(),
	exVolume: z.number().min(0).max(5).optional(),
	slideSe: z.number().min(0).max(49).optional(),
	slideVolume: z.number().min(0).max(5).optional(),
	touchHoldVolume: z.number().min(0).max(5).optional(),
	damageSeVolume: z.number().min(0).max(5).optional(),
	headPhoneVolume: z.number().min(0).max(10).optional(),
	sortTab: z.number().min(0).max(5).optional(),
	sortMusic: z.number().min(0).max(5).optional(),
	outFrameType: z.number().min(0).max(7).optional(),
	breakSlideVolume: z.number().min(0).max(5).optional(),
	touchVolume: z.number().min(0).max(5).optional()
});

export const UpdateMaiUserDataDto = z.strictObject({
	userName: z.string().min(1).regex(/^[Ａ-Ｚａ-ｚ０-９ぁ-んァ-ヶー一-龠々・．：；？！～／＋－×÷＝♂♀∀＃＆＊＠☆○◎◇□△▽♪†‡ΣαβγθφψωДё]+$/).optional(),
	iconId: z.number().min(0).optional(),
	plateId: z.number().min(0).optional(),
	titleId: z.number().min(0).optional(),
	partnerId: z.number().min(0).optional(),
	frameId: z.number().min(0).optional(),
	selectMapId: z.number().min(0).optional(),
	version: z.number().min(0),
});